model/FastViT/FasterViT-main/requirements.txt
timm~=1.0.7
tensorboardX==2.6.2.2
onnx==1.14.1
onnx_graphsurgeon==0.3.27
onnxruntime==1.15.1
polygraphy==0.47.1

matplotlib~=3.9.1
torchvision~=0.17.1
fastervit~=0.9.8
pillow~=10.2.0
setuptools~=68.2.2
numpy~=1.26.4
yaml~=0.2.5
pyyaml~=6.0.1
tensorboard~=2.17.0
tqdm~=4.66.4
scipy~=1.14.0